Improve compatibility with Paint.NET 4.2+
Added a constructor that uses the FileTypeOptions overload.
Changed the metadata handling to use the new Metadata class methods.

The new metadata class methods allow the Interoperability IFD and
other newer EXIF and GPS metadata that WIC supports to be preserved.

The Paint.NET version specific code has been placed into preprocessor
blocks, this allows the plugin to be built with support for either
Paint.NET 3.5.X or 4.2+.