Attempt to detect files with the wrong extension
It is fairly common to encounter JPEG/PNG images that have a WEBP file
extension.
The plugin will now check for the file signatures of a few common
formats when the image does not have a WebP file signature.

If a supported image format is detected, the image will be loaded using
GDI+.