Make lossless mode always use a quality value of 100
In lossless mode libwepb uses the quality value to control the
compression speed, with a value of 100 producing the smallest files.

I have received several reports from users that were confused by the
quality slider affecting the file size of an image that is saved using
lossless compression.

In version 1.3.12 and earlier lossless compression would always be used
with a quality value of 100.